`@remotion/studio`: Split selected outline overlay
